+++ app/devices/routes.py
@@ -126,6 +126,9 @@ def register():
         if Device.query.filter_by(mac_address=mac_address).first():
             flask.flash(_("Cet appareil est déjà enregistré !"), "danger")
         else:
+            if not rezident.devices:
+                # First device: create first connexion subscription
+                rezident.add_first_subscription()
             now = datetime.datetime.now(datetime.timezone.utc)
             device = Device(
                 rezident=rezident, name=form.nom.data,